requirements.txt
# AI Chatbot Testing Automation - Python Dependencies

# Core HTTP library for API requests
requests>=2.28.0

# OpenAI API client for advanced evaluation (optional)
openai>=1.0.0

# Google Gemini API client for question generation (optional)
google-genai>=0.3.0

# Standard libraries (included with Python 3.9+)
# json, os, argparse, logging, time, re, random, csv, datetime
# typing, pathlib, dataclasses, concurrent.futures

# Note: Install with: pip install -r requirements.txt
# For virtual environment: python3 -m venv venv && source venv/bin/activate && pip install -r requirements.txt
